pyproject.toml
[project]
name = "clawgovern"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.scripts]
clawgovern = "src.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.uv]
package = true
